+++ plugin/android/src/main/kotlin/ai/passio/nutrition_ai/NutritionAIHandler.kt
@@ -4,21 +4,30 @@ import ai.passio.nutrition_ai.converter.*
 import ai.passio.passiosdk.core.config.Bridge
 import ai.passio.passiosdk.core.config.PassioConfiguration
 import ai.passio.passiosdk.core.config.PassioMode
+import ai.passio.passiosdk.core.config.PassioStatus
 import ai.passio.passiosdk.passiofood.Barcode
 import ai.passio.passiosdk.passiofood.FoodCandidates
 import ai.passio.passiosdk.passiofood.FoodRecognitionListener
 import ai.passio.passiosdk.passiofood.PackagedFoodCode
 import ai.passio.passiosdk.passiofood.PassioID
 import ai.passio.passiosdk.passiofood.PassioSDK
+import ai.passio.passiosdk.passiofood.PassioStatusListener
 import ai.passio.passiosdk.passiofood.nutritionfacts.PassioNutritionFacts
 import android.app.Activity
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.graphics.drawable.BitmapDrawable
 import android.util.Log
+import android.net.Uri
 import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.launch
 import java.io.ByteArrayOutputStream
 
 class NutritionAIHandler(
@@ -40,6 +49,7 @@ class NutritionAIHandler(
             "fetchTagsFor" -> fetchTagsFor(call.arguments as String, result)
             "iconURLFor" -> fetchURLFor(call.arguments as HashMap<String, Any>, result)
             "transformCGRectForm" -> transformRect(call.arguments as HashMap<String, Any>, result)
+            "fetchNutrientsFor" -> fetchNutrientsFor(call.arguments as String, result)
         }
     }
 
@@ -139,6 +149,8 @@ class NutritionAIHandler(
         val argMap = arguments as HashMap<String, Any>
         when (argMap["method"]) {
             "startFoodDetection" -> startFoodDetection(argMap["args"] as Map<String, Any>, events)
+            // Sets up the PassioSDK status listener.
+            "setPassioStatusListener" -> setPassioStatusListener(events)
         }
     }
 
@@ -147,6 +159,8 @@ class NutritionAIHandler(
         val argMap = arguments as HashMap<String, Any>
         when (argMap["method"]) {
             "startFoodDetection" -> PassioSDK.instance.stopFoodDetection()
+            // Removes the PassioSDK status listener to stop receiving status updates.
+            "setPassioStatusListener" -> PassioSDK.instance.setPassioStatusListener(null)
         }
     }
 
@@ -158,7 +172,15 @@ class NutritionAIHandler(
                 image: Bitmap?,
                 nutritionFacts: PassioNutritionFacts?
             ) {
-                events.success(mapFromFoodCandidates(candidates))
+                // Use the Main dispatcher to launch a coroutine within the UI context
+                CoroutineScope(Dispatchers.Main).launch {
+                    // Call the detectionFlow function, passing in FoodCandidates and Bitmap image
+                    detectionFlow(candidates, image).collect {
+                        // Inside the collect block, which is a suspending function,
+                        // Send the recognition results through the event sink
+                        events.success(it)
+                    }
+                }
             }
         }, config)
     }
@@ -262,5 +284,101 @@ class NutritionAIHandler(
                 resultRect.height().toDouble()
             )
         )
+    }
+
+    /**
+     * Fetches nutrients for a given passioID and communicates the result through the provided callback.
+     *
+     * @param args A HashMap containing method arguments, with "passioID" as a required key.
+     * @param callback A MethodChannel.Result used to communicate the result back to the caller.
+     */
+    private fun fetchNutrientsFor(passioID: PassioID, callback: MethodChannel.Result) {
+        // Calling PassioSDK to fetch nutrients for the given passioID
+        PassioSDK.instance.fetchNutrientsFor(passioID) { list ->
+            // Mapping PassioNutrient objects to a new list using mapFromPassioNutrient function
+            val nutrientList = list?.map { mapFromPassioNutrient(it) }
+            // Calling the callback's success method with the nutrientList
+            callback.success(nutrientList)
+        };
+    }
+
+    // Define a suspend function named detectionFlow that takes FoodCandidates and a Bitmap image as parameters
+    suspend fun detectionFlow(candidates: FoodCandidates, image: Bitmap?) = flow {
+
+        // Creating a Map from the FoodCandidates object
+        val mapCandidates = mapFromFoodCandidates(candidates)
+
+        // Creating a Map from the image object
+        val imageMap = mapFromBitmap(image)
+
+        // Creating a mutable map to store results
+        val resultMap = mutableMapOf<String, Any?>()
+        // Adding the map of FoodCandidates to the resultMap
+        resultMap["candidates"] = mapCandidates
+        // Adding the map representing image properties to the resultMap
+        resultMap["image"] = imageMap
+
+        // Emit the resultMap as a flow item
+        emit(resultMap)
+    }.flowOn(Dispatchers.IO) // Specify that the flow should run on the IO dispatcher
+
+
+    /**
+     * Sets up the PassioSDK status listener for handling various events related to file downloads and Passio status changes.
+     *
+     * @param events The EventSink to send Flutter events to.
+     */
+    private fun setPassioStatusListener(events: EventChannel.EventSink) {
+        PassioSDK.instance.setPassioStatusListener(object : PassioStatusListener {
+
+            /**
+             * Called when all files have been successfully downloaded.
+             *
+             * @param fileUris List of URIs for the downloaded files.
+             */
+            override fun onCompletedDownloadingAllFiles(fileUris: List<Uri>) {
+                val files = fileUris.map(Uri::toString)
+                val statusListenerMap = mapFromPassioStatusListener(
+                    "completedDownloadingAllFiles",
+                    files
+                )
+                events.success(statusListenerMap)
+            }
+
+            /**
+             * Called when an individual file has been successfully downloaded.
+             *
+             * @param fileUri URI of the downloaded file.
+             * @param filesLeft Number of files left to download.
+             */
+            override fun onCompletedDownloadingFile(fileUri: Uri, filesLeft: Int) {
+                val downloadingMap = mapFromCompletedDownloadingFile(fileUri, filesLeft)
+                val statusListenerMap =
+                    mapFromPassioStatusListener("completedDownloadingFile", downloadingMap)
+                events.success(statusListenerMap)
+            }
+
+            /**
+             * Called when an error occurs during the download process.
+             *
+             * @param message Error message describing the issue.
+             */
+            override fun onDownloadError(message: String) {
+                events.success(mapFromPassioStatusListener("downloadingError", message))
+            }
+
+            /**
+             * Called when the Passio status changes.
+             *
+             * @param status The new PassioStatus.
+             */
+            override fun onPassioStatusChanged(status: PassioStatus) {
+                val passioStatus = mapFromPassioStatus(status)
+                val statusListenerMap =
+                    mapFromPassioStatusListener("passioStatusChanged", passioStatus)
+                events.success(statusListenerMap)
+            }
+
+        });
     }
 }