+++ plugin/android/src/main/kotlin/ai/passio/nutrition_ai/converter/OutputConverter.kt
@@ -1,6 +1,7 @@
 package ai.passio.nutrition_ai.converter
 
 import ai.passio.nutrition_ai.R
+import ai.passio.nutrition_ai.utils.toByteArray
 import ai.passio.passiosdk.core.config.PassioMode
 import ai.passio.passiosdk.core.config.PassioSDKError
 import ai.passio.passiosdk.core.config.PassioStatus
@@ -9,6 +10,7 @@ import ai.passio.passiosdk.passiofood.DetectedCandidate
 import ai.passio.passiosdk.passiofood.FoodCandidates
 import ai.passio.passiosdk.passiofood.PackagedFoodCandidate
 import ai.passio.passiosdk.passiofood.PassioID
+import ai.passio.passiosdk.passiofood.PassioNutrient
 import ai.passio.passiosdk.passiofood.data.measurement.Grams
 import ai.passio.passiosdk.passiofood.data.measurement.KiloCalories
 import ai.passio.passiosdk.passiofood.data.measurement.Kilograms
@@ -27,6 +29,8 @@ import ai.passio.passiosdk.passiofood.data.model.PassioIDEntityType
 import ai.passio.passiosdk.passiofood.data.model.PassioServingSize
 import ai.passio.passiosdk.passiofood.data.model.PassioServingUnit
 import android.content.Context
+import android.graphics.Bitmap
+import android.net.Uri
 
 fun mapFromPassioStatus(status: PassioStatus): Map<String, Any?> {
     val map = mutableMapOf<String, Any?>()
@@ -178,6 +182,8 @@ fun mapFromFoodOrigin(origin: PassioFoodOrigin): Map<String, Any?> {
 
 fun mapFromFoodItemData(data: PassioFoodItemData): Map<String, Any?> {
     val dataMap = mutableMapOf<String, Any?>()
+    val originalQuantity = data.selectedQuantity
+    val originalUnit = data.selectedUnit
     dataMap["passioID"] = data.passioID
     dataMap["name"] = data.name
     dataMap["selectedQuantity"] = data.selectedQuantity
@@ -224,6 +230,7 @@ fun mapFromFoodItemData(data: PassioFoodItemData): Map<String, Any?> {
     dataMap["magnesium"] = mapFromUnitMass(data.totalMagnesium())
     dataMap["phosphorus"] = mapFromUnitMass(data.totalPhosphorus())
     dataMap["iodine"] = mapFromUnitMass(data.totalIodine())
+    data.setServingSize(originalUnit, originalQuantity)
     return dataMap
 }
 
@@ -257,4 +264,72 @@ fun mapFromSearchResult(result: Pair<PassioID, String>): Map<String, Any?> {
     searchMap["passioID"] = result.first
     searchMap["name"] = result.second
     return searchMap
+}
+
+fun mapFromBitmap(image: Bitmap?): MutableMap<String, Any?>? {
+    // Checking if the 'image' is not null before performing operations
+    val byteArray: ByteArray? = image?.toByteArray()
+
+    // Creating a map from image and byteArray properties if 'image' is not null
+    return image?.let {
+        // Initializing a mutable map to store image properties
+        val map = mutableMapOf<String, Any?>()
+
+        // Adding width, height, and pixel data to the map
+        map["width"] = it.width
+        map["height"] = it.height
+        map["pixels"] = byteArray!!
+
+        // The resulting map containing image properties
+        map
+    }
+}
+
+/// Maps the information from a completed downloading file to a [Map].
+///
+/// Parameters:
+/// - [fileUri]: The URI of the completed file.
+/// - [filesLeft]: The number of files left to download.
+///
+/// Returns a [Map] containing the mapped information.
+fun mapFromCompletedDownloadingFile(fileUri: Uri, filesLeft: Int): Map<String, Any?> {
+    val downloadMap = mutableMapOf<String, Any?>()
+    downloadMap["fileUri"] = fileUri.path
+    downloadMap["filesLeft"] = filesLeft
+    return downloadMap
+}
+
+/**
+ * This function takes an `event` and `data` as parameters and maps them into a
+ * `Map` representing the structure of PassioStatusListener events.
+ *
+ * @param event The type of event (e.g., "onPassioStatusChanged").
+ * @param data The data associated with the event. It can be of any type.
+ * @return A `Map` containing the event type and associated data.
+ */
+fun mapFromPassioStatusListener(event: String, data: Any?): Map<String, Any?> {
+    val statusMap = mutableMapOf<String, Any?>()
+    statusMap["event"] = event
+    statusMap["data"] = data
+    return statusMap
+}
+
+/**
+ * Maps a [PassioNutrient] object to a [Map] for serialization.
+ *
+ * @param nutrient The [PassioNutrient] object to be mapped.
+ * @return A [Map] containing the serialized representation of the [PassioNutrient].
+ */
+fun mapFromPassioNutrient(nutrient: PassioNutrient): Map<String, Any?> {
+    // Create a mutable map to store the serialized nutrient data.
+    val nutrientMap = mutableMapOf<String, Any?>()
+
+    // Add nutrient properties to the map.
+    nutrientMap["amount"] = nutrient.amount
+    nutrientMap["inflammatoryEffectScore"] = nutrient.inflammatoryEffectScore
+    nutrientMap["name"] = nutrient.name
+    nutrientMap["unit"] = nutrient.unit
+
+    // Return the serialized nutrient map.
+    return nutrientMap
 }