+++ plugin/example/lib/util/double_extension.dart
@@ -1,3 +1,3 @@
 extension Round on double {
-  int get roundUpAbs => this.isNegative ? this.floor() : this.ceil();
+  int get roundUpAbs => isNegative ? floor() : ceil();
 }