+++ plugin/ios/Classes/view/PassioPreview.swift
@@ -42,4 +42,8 @@ class PassioPreview: NSObject, FlutterPlatformView {
         }
     }
     
+    deinit {
+        PassioNutritionAI.shared.removeVideoLayer()
+    }
+
 }