+++ plugin/lib/src/converter/platform_input_converter.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import '../nutrition_ai_detection.dart';
 import '../nutrition_ai_configuration.dart';
+import '../models/nutrition_ai_nutrient.dart';
 
 Map<String, dynamic> mapOfPassioConfiguration(PassioConfiguration config) {
   return {
@@ -32,5 +33,23 @@ Map<String, double> mapRectangle(Rectangle<double> rectangle) {
     'top': rectangle.top,
     'width': rectangle.width,
     'height': rectangle.height
+  };
+}
+
+/// Converts a [PassioNutrient] object to a [Map].
+///
+/// This function takes a [PassioNutrient] instance and converts it into a [Map]
+/// with keys representing nutrient properties and corresponding values.
+///
+/// Parameters:
+/// - [nutrient]: The [PassioNutrient] object to be mapped.
+///
+/// Returns a [Map] containing nutrient information.
+Map<String, dynamic> mapOfPassioNutrient(PassioNutrient nutrient) {
+  return {
+    'amount': nutrient.amount,
+    'inflammatoryEffectScore': nutrient.inflammatoryEffectScore,
+    'name': nutrient.name,
+    'unit': nutrient.unit,
   };
 }